requirements.txt
PyYAML==3.12
psutil=5.4.3
